+++ Material.Blazor/Model/Snackbar/MBSnackbarServiceConfiguration.cs
@@ -12,38 +12,43 @@ namespace Material.Blazor
         public const bool DefaultLeading = false;
         public const bool DefaultStacked = false;
         public const bool DefaultDismissIcon = true;
+        public const MBNotifierCloseMethod DefaultCloseMethod = MBNotifierCloseMethod.TimeoutAndDismissButton;
+
+
+        private MBNotifierCloseMethod closeMethod = DefaultCloseMethod;
+        /// <summary>
+        /// Determines how the snackbar closes. Defaults to <see cref="MBNotifierCloseMethod.TimeoutAndDismissButton"/>.
+        /// </summary>
+        public MBNotifierCloseMethod CloseMethod { get => closeMethod; set => Setter(ref closeMethod, value); }
 
 
         private int? timeout = DefaultTimeout;
-
         /// <summary>
         /// The snackbar's timeout in milliseconds.
         /// Minimum value is 4000 (4 seconds), maximum value is 10000 (10s).
-        /// Use -1 to disable.
         /// Defaults to 5000.
         /// </summary>
         public int? Timeout
         {
             get => timeout;
-            set => Setter(ref timeout, value switch { null => null, -1 => -1, _ => Math.Max(4000, Math.Min(value.Value, 10000)) });
+            set => Setter(ref timeout, value switch { null => null, _ => Math.Max(4000, Math.Min(value.Value, 10000)) });
         }
+
+
         /// <summary>
         /// By default, snackbars are centered horizontally within the viewport.
         /// On larger screens, they can optionally be displayed on the leading edge of the screen (the left side in LTR, or the right side in RTL) by setting this property to true.
         /// Default: false.
         /// </summary>
         public bool? Leading { get; set; }
+
+
         /// <summary>
         /// Action buttons with long text should be positioned below the label instead of alongside it.
         /// This can be accomplished by setting this property to true.
         /// Default: false.
         /// </summary>
         public bool? Stacked { get; set; }
-        /// <summary>
-        /// Snackbars are intended to dismiss on their own after a few seconds, but a dedicated dismiss icon may be optionally included as well for accessibility purposes.
-        /// By default, a dismiss icon is displayed. This can be disabled by setting this property to false.
-        /// </summary>
-        public bool? DismissIcon { get; set; }
 
 
         /// <summary>