+++ Material.Blazor/Model/Toast/IMBToastService.cs
@@ -4,18 +4,18 @@ namespace Material.Blazor
 {
     /// <summary>
     /// Interface for the Material.Blazor toast service, developed from the code base of Blazored Toast by Chris Sainty.
-    /// Works in conjunction with a <see cref="MBToastAnchor"/> that must be placed in either App.razor or
+    /// Works in conjunction with a <see cref="MBAnchor"/> that must be placed in either App.razor or
     /// MainLayout.razor to avoid an exception being thrown when you first attempt to show a toast notification.
     /// 
     /// <para>
     /// Throws a <see cref="System.InvalidOperationException"/> if
-    /// <see cref="ShowToast(MBToastLevel, string, string, MBToastCloseMethod?, string, string, IMBIconFoundry?, bool?, uint?, bool)"/>
-    /// is called without an <see cref="MBToastAnchor"/> component used in the app.
+    /// <see cref="ShowToast(MBToastLevel, string, string, MBNotifierCloseMethod?, string, string, IMBIconFoundry?, bool?, uint?, bool)"/>
+    /// is called without an <see cref="MBAnchor"/> component used in the app.
     /// </para>
     /// <example>
     /// <para>You can optionally add configuration when you add this to the service collection:</para>
     /// <code>
-    /// services.AddMBToastService(new MBToastServiceConfiguration()
+    /// services.AddMBServices(toastServiceConfiguration: new MBToastServiceConfiguration()
     /// {
     ///     Postion = MBToastPosition.TopRight,
     ///     CloseMethod = MBToastCloseMethod.Timeout,
@@ -68,7 +68,7 @@ namespace Material.Blazor
             MBToastLevel level,
             string message,
             string heading = null,
-            MBToastCloseMethod? closeMethod = null,
+            MBNotifierCloseMethod? closeMethod = null,
             string cssClass = null,
             string iconName = null,
             IMBIconFoundry? iconFoundry = null,